d6 p2

+++ src/d6/Part1.java
@@ -8,15 +8,13 @@ public class Part1 {
     public static void main(String[] args) throws IOException {
         List<Integer> times = List.of(53, 91, 67, 68);
         List<Integer> distances = List.of(250, 1330, 1081, 1025);
-//        List<Integer> times = List.of(7, 15, 30);
-//        List<Integer> distances = List.of(9, 40, 200);
         int totalWinFactor = 1;
 
         for (int i = 0; i < times.size(); i++) {
             int wins = 0;
             for (int j = 0; j < times.get(i); j++) {
                 Race race = new Race(j, times.get(i) - j);
-                int distance = race.getDistance();
+                long distance = race.getDistance();
                 if (distance > distances.get(i)) {
                     wins++;
                 }
